Moves diagnostics set up to diagnostics module
